chronos
git-svn-id: svn://elaine/murooma/trunk@290 d8a302eb-03bc-478d-80e4-98257eca68ef

+++ message/sampleFormat.h
@@ -22,9 +22,9 @@ public:
 	uint16_t sampleSize;
 	uint16_t frameSize;
 
-	float msRate() const
+	double msRate() const
 	{
-		return (float)rate/1000.f;
+		return (double)rate/1000.f;
 	}
 
 	virtual void read(std::istream& stream)
